factor out sending ACKs and NACKs into their own functions
also made some changes to the generated protobufs - acknowledgement is
now its own variant of the Routing packet as opposed to the
special-cased NO_ERROR value in the error variant of the packet.

None of this changes the communication between the nodes yet - nodes on
this firmware communicate as usual just with different underlying types
however this change and future ones are incompatible with the standard
meshtastic firmware.